requirements.txt
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
fastapi>=0.135.0,<0.136.0
uvicorn[standard]>=0.43.0,<0.44.0
python-dotenv>=1.2.1
PyJWT>=2.10.1
python-multipart>=0.0.26
openrouter>=0.1.0